backend/requirements.txt
fastapi
sqlalchemy
psycopg2-binary
python-dotenv
uvicorn
requests
